% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCoordinates.R
\name{extractCoordinates}
\alias{extractCoordinates}
\title{Extrcat coordinates by coordinates}
\usage{
extractCoordinates(wholeCoordinates)
}
\arguments{
\item{wholeCoordinate}{the latitude and longitude separated by a white space}
}
\value{
A list which contains lat and long as attributes
}
\description{
Extrcat coordinates by coordinates
}
