% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCoordinates.R
\name{extractCoordinatesByLine}
\alias{extractCoordinatesByLine}
\title{Extract coordinates by line}
\usage{
extractCoordinatesByLine(line)
}
\arguments{
\item{line}{the line to extract coordinates from}
}
\value{
A list which contains lat and long as attributes
}
\description{
Extract coordinates by line
}
