% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractInfo.R
\name{extractInfo}
\alias{extractInfo}
\title{Extracting information from a single line of data}
\usage{
extractInfo(line)
}
\arguments{
\item{line}{A single line of data which to extract info}
}
\value{
Extracted data as a list
}
\description{
Extracting information from a single line of data
}
\examples{
data <- c();
addAgent("driver", "circle", "phase 1", "42.23 12.23", "no path", "blue", data)
extractInfo(data[1])
}
