% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractInfo.R
\name{extractInfoByIndex}
\alias{extractInfoByIndex}
\title{Extracting information from a single line of data by index}
\usage{
extractInfoByIndex(data, agentType, agentName)

extractInfoByIndex(data, agentType, agentName)
}
\arguments{
\item{data}{The data to extract the information from}

\item{agentType}{agentType of the line supposed to be extracted}

\item{agentName}{agentName of the line supposed to be extracted}

\item{index}{The index of the line in the data array to extract.}
}
\value{
Extracted data as a list

Extracted data as a list
}
\description{
Extracting information from a single line of data by index

Extracting information from a single line of data by agentType and agentName
}
\examples{
data <- c();
addAgent("driver", "circle", "phase 1", "42.23 12.23", "no path", "blue", data)
extractInfoByIndex(data, 1)
}
