% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{openMap}
\alias{openMap}
\title{Opening the map for graphically accesses}
\usage{
openMap(data)
}
\arguments{
\item{data}{The data which contains line of information}
}
\description{
Opening the map for graphically accesses
}
