% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{Loading ABM data file}
\usage{
readData(path)
}
\arguments{
\item{path}{The file path}
}
\value{
A list of files' lines
}
\description{
Loading ABM data file
}
\examples{
data <- readData(myPath)
data[5]
}
