% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeAgent.R
\name{removeAgentByName}
\alias{removeAgentByName}
\title{Remove Agent}
\usage{
removeAgentByName(agentName, data)
}
\arguments{
\item{agentName}{The name of the agent to be removed}

\item{data}{The data which contains line of the output}
}
\value{
The data which has been adjusted
}
\description{
Remove Agent
}
