% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeAgent.R
\name{removeAgentByStartCoordinate}
\alias{removeAgentByStartCoordinate}
\title{Remove Agent}
\usage{
removeAgentByStartCoordinate(startCoordinate, data)
}
\arguments{
\item{startCoordinate}{The coordinate's of the location to start.}

\item{data}{The data which contains line of the output}
}
\value{
The data which has been adjusted
}
\description{
Remove Agent
}
